refactor(examples): changed the example list structure
