+++ src/features/query/ExamplesButton.jsx
@@ -2,19 +2,19 @@ import React, { useState } from 'react';
 import { Tooltip } from 'flowbite-react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-import Example from './Example';
+import ExampleList from './ExampleList';
 import Modal from '@/components/Modal';
-
-import { examples } from '@/config';
 
 export default function ExamplesButton () {
   const [modalOpen, setModalOpen] = useState(false);
+
+  const showModal = show => () => setModalOpen(show);
 
   return (
     <>
       <button
         className='!px-6 cta-blue-wired font-montserrat font-medium flex gap-2'
-        onClick={() => setModalOpen(true)}
+        onClick={showModal(true)}
       >
         <span>Exemples</span>
         <div className='m-auto'>
@@ -25,20 +25,14 @@ export default function ExamplesButton () {
       </button>
 
       {modalOpen && (
-        <Modal onClose={() => setModalOpen(false)}>
+        <Modal onClose={showModal(false)}>
           <Modal.Header>Exemples de requêtes</Modal.Header>
           <Modal.Body>
-            <div className='p-1 flex flex-col gap-2'>
+            <div className='p-1 flex flex-col gap-5'>
               <p className='max-w-2xl'>
                 Voici quelques exemples dont vous pouvez vous inspirer pour votre recherche. Cliquez sur l'une des loupes et la zone de requête sera remplie automatiquement par le contenu de l'exemple choisi. Cet échantillon illustre différentes façons d'interroger l'API Istex en utilisant :
               </p>
-              <div>
-                {examples.map(example => (
-                  <div key={example.label}>
-                    <Example info={example} onClose={() => setModalOpen(false)} />
-                  </div>
-                ))}
-              </div>
+              <ExampleList onClose={showModal(false)} />
             </div>
           </Modal.Body>
           <Modal.Footer />
